requirements.txt
pyautogui
keyboard
pywin32
pyperclip